// project/src/models/enums/EquipmentSlots.ts
export enum EquipmentSlots
    {
    HEADWEAR = "Headwear",
    EARPIECE = "Earpiece",
    FACE_COVER = "FaceCover",
    ARMOR_VEST = "ArmorVest",
    EYEWEAR = "Eyewear",
    ARM_BAND = "ArmBand",
    TACTICAL_VEST = "TacticalVest",
    POCKETS = "Pockets",
    BACKPACK = "Backpack",
    SECURED_CONTAINER = "SecuredContainer",
    FIRST_PRIMARY_WEAPON = "FirstPrimaryWeapon",
    SECOND_PRIMARY_WEAPON = "SecondPrimaryWeapon",
    HOLSTER = "Holster",
    SCABBARD = "Scabbard"
}